+++ lib_spdif/api/spdif.h
@@ -6,7 +6,7 @@
 #include <stddef.h>
 #include <xs1.h>
 #include <xccompat.h>
-#if defined(__DOXYGEN__) || defined(__XC__) 
+#if defined(__DOXYGEN__) || defined(__XC__)
 /* These are needed to allow doxy to render properly */
 #define in_port_t in port
 #define out_port_t out port
@@ -41,6 +41,22 @@
 
 /* Helper macro for extracting sample bits from received S/PDIF subframe */
 #define SPDIF_RX_EXTRACT_SAMPLE(x) ((x & 0xFFFFFFF0) << 4)
+
+/**
+ * @def SPDIF_TX_ENABLE_PRO_CHANNEL_STATUS
+ *
+ * @brief Selects whether the S/PDIF transmitter builds a
+ *        Professional (AES3) channel status block or a
+ *        Consumer (S/PDIF) channel status block.
+ *
+ * Building consumer block by default.
+ * Can be overriden by defining SPDIF_TX_ENABLE_PRO_CHANNEL_STATUS (1)
+ * or passing a compiler build flag -DSPDIF_TX_ENABLE_PRO_CHANNEL_STATUS=1
+ * to build professional channel status block.
+ */
+#ifndef SPDIF_TX_ENABLE_PRO_CHANNEL_STATUS
+#define SPDIF_TX_ENABLE_PRO_CHANNEL_STATUS (0)
+#endif
 
 /** S/PDIF receive function.
  *